requirements.txt
fastapi
uvicorn
pydantic
openai
requests    
python-dotenv